Refactored routes tests into seperate files
